Rename Method + Resolving
Rename method from "trace" to "send_range". The "trace" naming may be misleading as it contains no special logic for a destination response. This method simply sends packets for all hops in the range within the timeout.

Ekko no longer resolves domain names, bumping minor version number as this is a breaking change.